Skip already translated files

+++ translate.py
@@ -12,13 +12,16 @@ def translate(data_path: Path, src_lang: str, dest_lang: str):
     dest_path.mkdir(parents=True, exist_ok=True)
 
     for src_path in tqdm(list(data_path.glob(f"{src_lang}/*.json"))):
+        output_path = dest_path / src_path.name
+        if output_path.exists():
+            continue
         with open(src_path, "r", encoding="utf-8") as f:
             src_data = json.load(f)
         dest_data = {}
         translations = translator.translate(list(src_data.values()), src=src_lang, dest=dest_lang)
         for key, translation in zip(src_data.keys(), translations):
             dest_data[key] = translation.text
-        with open(dest_path / src_path.name, "w", encoding="utf-8") as f:
+        with open(output_path, "w", encoding="utf-8") as f:
             json.dump(dest_data, f)
 
 
@@ -27,8 +30,6 @@ def main():
     SRC_LANG = "zh-tw"
     for dest_lang in tqdm(LANGUAGES):
         if dest_lang == SRC_LANG:
-            continue
-        if dest_lang in {"en", "vi"}:  # TODO: remove this
             continue
         translate(DATA_PATH, SRC_LANG, dest_lang)
